Fix python version in CI
The environment action was not using the python-version input variable
correctly. This caused the latest Python to be installed in some cases,
instead of the indicated Python. Python 3.11 CI builds are currently
broken, causing the entire pipeline to fail.

Fixing the action means it will build on Python 3.10, which will
(hopefully) still pass CI. Fixing CI on Python 3.11 will come later.