Handle latitude/longitude bounds properly when making polygons
Previously the bounds were ignored, with polygons being constructed by
finding midpoints between face centres. This approach is still used as
a fallback, but if a dataset defines bounds for a coordinate, those
bounds are used to construct the polygons.

This means CF grid topologies with 2D coordinates can have properly
defined edge cells, if the bounds are defined in the dataset.