Rename LICENCE to LICENSE
When building the source tarball the LICENCE file is correctly included,
but somewhere along the way the file is dropped and not included in the
wheel. This is not great already, but then the conda-forge package fails
to build because there is no LICENSE file.

The easiest solution is to spell licence incorrectly.

The yanks win this one...