Use ravel/wind as consistent antonyms
Previously there was make_linear + ravel_index for straightening
something, and unravel_index for transforming back to
a multi-dimensional structure.

Ravel and unravel mean the same thing as verbs, which is confusing.
Ravel was kept as the straightening verb to be consistent with
numpy.ravel(). Wind was chosen as the verb for the inverse operation.