Make coasts and gridlines optional on plots
There are currently a number of problems with gridlines on plots in
Jupyter notebooks. There are different solutions that depend on whether
you want an interactive plot, a static plot in a notebook, or are
saving the plot to a file. Unfortunately there is no single solution
that works in all cases currently. Until a proper solution is found,
users can disable gridlines and then re-enable them in the way that
works for their current environment.

https://github.com/SciTools/cartopy/issues/2245
https://github.com/SciTools/cartopy/issues/2246
https://github.com/SciTools/cartopy/issues/2247