Fix some warnings raised in the tests with xarray 2023.09.0
xarray now explicitly complains if you're doing Bad Things, instead of
silently trying to work with it. Explicit warnings are good! But now
I have to not do those Bad Things.

This has the upside of removing one of the warnings the tests silenced,
which can only be a good thing.