Check dataset polygons are valid when constructing them
`Convention.polygons` is now defined on the base `Convention` class. It
calls a new abstract method `Convention._make_polygons()`. This allows
us to do some post processing of the polygons as part of generating
them.

Invalid polygons generate a warning and are dropped as part of
generation. Non-simple polygons generate a warning but are kept.

This means the triangulation code doesn't need to perform this step. It
doesn't save any time overall, as the polygons still need to be tested
either way.