Experimental support for plotting SCHISM output
SCHISM datasets define many variables on the node grid of the dataset,
unlike COMPAS datasets with their data on the face grid. emsarray can
open SCHISM datasets but didn't know how to plot them till now.

This support is experimental and the API is likely to change. There are
no tests yet.