Move more plotting logic back to the Convention
Now the convention can decide how to plot a variable / tuple of
variables, perhaps by determining which grid they are on. The collection
of new plot functions on the Convention also give more flexibility to
callers.

The animation functions have not been updated to match yet.

This is a partial implementation of the work required for #175