Fix Cartopy docs URL
The Cartopy docs are now on readthedocs. We use intersphinx for building
docs that link to other projects. As the docs have been moved and there
is no redirect for the `objects.inv` file our docs build was breaking.
Easily fixed by updating the URL for intersphinx to the new location.

See also: https://github.com/SciTools/cartopy/pull/2558